[ArmAutoSignoff] Split check into two parts (source code, labels+checks) (#32593)
