+++ .github/actions/add-label-artifact/action.yaml
@@ -13,15 +13,7 @@ runs:
   using: composite
 
   steps:
-    - name: Create empty file to upload artifact
-      run: "> $RUNNER_TEMP/empty.txt"
-      shell: bash
-
-    # The maximum length is reported to be 260 characters. A full list of invalid artifact name characters is documented here:
-    # https://github.com/actions/toolkit/blob/main/packages/artifact/src/internal/upload/path-and-artifact-name-validation.ts
-    - uses: actions/upload-artifact@v4
+    - uses: ./.github/actions/add-empty-artifact
       with:
-        name: label-${{ inputs.name }}=${{ inputs.value }}
-        path: ${{ runner.temp }}/empty.txt
-        if-no-files-found: error
-        overwrite: true
+        name: label-${{ inputs.name }}
+        value: ${{ inputs.value }}
